syft: move resolve pointer type reply to messages and add docstrings
